% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHCDS.R
\name{chcds}
\alias{chcds}
\title{Conformalizes conditional highest density sets with an additive conformal adjustment}
\usage{
chcds(haty_cal, haty_grid, y_grid, alpha)
}
\arguments{
\item{haty_cal}{the vector of estimated density values for response values observed in the calibration set}

\item{haty_grid}{the matrix of estimated density values over the response range for the calibration and test sets}

\item{y_grid}{the vector containing grided values of y over the response range}

\item{alpha}{numeric; the miscoverage rate}
}
\value{
\if{html}{\out{<div class="sourceCode">}}\preformatted{A numeric matrix of the conformal prediction sets, up to the union of three sets. If the returned set is less than the union of three sets, NA values are returned for additional sets.
}\if{html}{\out{</div>}}
}
\description{
Conformalized estimated highest conditional density sets with an additive conformal adjustments
}
\examples{
set.seed(1)
x <- runif(500)
y <- rnorm(500, 2 * x, 1)
mod <- lm(y ~ x)
x_cal <- runif(500)
y_cal <- rnorm(500, 2 * x, 1)
x_test <- runif(5) ## 5 test points
preds_cal <- predict(mod, newdata = data.frame(x = x_cal))
preds_test <- predict(mod, newdata = data.frame(x = x_test))
norm_means <- c(preds_cal, preds_test)
haty_cal <- dnorm(y_cal, preds_cal, sigma(mod)) ## using a Normal density
y_grid <- seq(min(y_cal) * 501 / 500, max(y_cal) * 501 / 500, length.out = 1000)
haty_grid <- matrix(NA, nrow = 505, ncol = 1000)
for(ii in seq(505)){
    haty_grid[ii, ] <- dnorm(y_grid, norm_means[ii], sigma(mod))
}
alpha <- 0.20 ## 80\\% prediction intervals
chcds(haty_cal, haty_grid, y_grid, alpha)
}
